Feature/backend/authorization (#5)
User-based authorization for private endpoint support

This feature supports the implementation of private endpoints by enabling a user based authentication scheme. This allows further work with implementing data entry APIs.

* Added database classes and data properties for user authorization

Some database modification was necessary in order to support authorization / authentication because we did not want to rely on the integrated database user management.

The main modification is the addition of a "User" class which represents all users. Users are expected to have a unique username data property (the uniqueness constraint is a todo), a passwordHash data property, a userRole object property, and a creation data property. One test user is already preinserted into the database, and additional users should be added under the "namespace" "sdg.user."

* Initial authorization / authentication implementation

The fundamental implementation is based on bcrypt for hashing passwords,
and JWT as a token implementation (using the "jsonwebtoken" library).
By using JWTs, we alleviate the need to access the database when
verifying if the token originated with us.

* Adding easy token verification.
Token validity can be easily checked using the 'verifyToken' routing middleware.

* Fix token verification and improve handling of api errors.

* Add functional property to user related data properties in database

* Improve post test for api

* Fix eslint errors in order to appease our machine overlords...

Co-authored-by: Fredrik Busklein <61384979+fredrbus@users.noreply.github.com>